+++ trl/trainer/base_trainer.py
@@ -58,6 +58,7 @@ _TELEMETRY_TRAINERS = {
     "PRMTrainer",
     "SDFTTrainer",
     "SDPOTrainer",
+    "ServerDistillationTrainer",
     "SSDTrainer",
     "TPOTrainer",
     "XPOTrainer",
